Recreated DNS server (back to catalog)
trust-dns 0.16 fixed the wildcard entry which didn't work in previous
versions and changed a lot of the low level implementation of Authority
which causes me to go back to Catalog implementation - hopefully there
will be less maintenance going forward.

The price is removing support for subdomains. Every required sub-domains
needs to be specified explicitly.