Migrated state to symlinks/text files on disk.
Instead of having some sort state db (in whatever format) and managing
state via webapp, I decided to have a much simpler state held in a
"state" directoty in the form of symlinks and text files which hold
configurations. This way one can manage the state manually and just
trigger a reload command in the duwopctl tool (not yet implemented).

This was a large(ish) change and caused a few fixes to issues raised by
the change.

Still a lot to do.