BREAKING CHANGE: changed proxy configuration.
A URL was not a good choice (although it matched the hyper_reverse_proxy
usage). In preparation to the new proxy, The only configuration required
is port. The IP part while exists is currently ignored. I might decide
to add later the option to proxy addresses other than localhost so I
added it to avoid another breaking change.