Reload SSL :)
This was really hard, It took me quite a while to find a way to
implement this elegently. The only issue is that the reload is async and
I don't know when running from the client if it succeeded.
Also, features (like notifying the user if ssl is disabled when running
reload-ssl) were skipped in order to keep the code simpler.